+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @ricsam/isolate-test-utils
 
+## 0.1.24
+
+### Patch Changes
+
+- Updated dependencies
+  - @ricsam/isolate-runtime@0.1.26
+  - @ricsam/isolate-core@0.1.19
+  - @ricsam/isolate-fetch@0.1.23
+  - @ricsam/isolate-fs@0.1.19
+
 ## 0.1.23
 
 ### Patch Changes